abfsClient.getBlobProperty to return abfsRestOperation object
